maskrcnn/requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
clearml==1.16.1
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fonttools==4.52.1
fsspec==2024.5.0
furl==2.1.3
fvcore==0.1.5.post20221221
grpcio==1.64.0
hydra-core==1.3.2
idna==3.7
iopath==0.1.9
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
orderedmultidict==1.0.1
packaging==24.0
parso==0.8.4
pathlib2==2.3.7.post1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wget==3.2
yacs==0.1.8
